+++ CHANGELOG.md
@@ -10,8 +10,8 @@
 * FilterChooser field suggestions now search within multi-word field names.
 * Autosize performance has been improved for very large grids.
 * `DashCanvas.extraMenuItems` and `DashContainer.extraMenuItems` are now observable properties
-* `ContextMenuItem.hidden` and `MenuItem.hidden` now additionally accept callback functions to
-  be evaluated at runtime when their container menu is rendered
+* `ContextMenuItem` and `MenuItem` now accept a `displayFn` callback for dynamic overriding
+  of an individual item's display configs upon render
 
 ### 🐞 Bug Fixes